Produce summary table of completeness and validity for variables in a dataframe

Runs completeness and validity calculations on input dataframe df. Filters data by reporter to reduce process time. Use with map() to run on multiple reporters

Usage

makeValidData(
  df,
  indexCol = "reporter",
  filtRep,
  rulesObj = NULL,
  ruleset = "default"
)

Arguments

df

A dataframe whose column values are subjected to validation rules via switch()

indexCol

The column to filter on. Should contain the values given in `filtRep`

filtRep

Character A length 1 vector of value to filter on before

rulesObj

List Named list of validation functions

ruleset

Character Grouping name of custom ruleset. calculation (e.g., input$filtRep a checkboxGroup of reporter names)

Value

Tibble with same columns as input DF, plus column "metric", and 2 rows: Completeness percentage and validity percentage

See also

Examples

df <- data.frame('myindex' = c('Bob','Bob','Larry'),
'lastname' = c('AAA','BBB','CCC'),
'gender' = c('Male','Female','M A L E'), 'norule' = c('A','B','C'))
makeValidData(df, indexCol = 'myindex', filtRep = 'Bob')
#> # A tibble: 2 × 5
#>   myindex metric       lastname gender norule
#>   <fct>   <chr>           <dbl>  <dbl>  <dbl>
#> 1 Bob     Completeness      100    100    100
#> 2 Bob     Validity          100    100     NA
makeValidData(df, indexCol = 'myindex', filtRep = 'Larry')
#> # A tibble: 2 × 5
#>   myindex metric       lastname gender norule
#>   <fct>   <chr>           <dbl>  <dbl>  <dbl>
#> 1 Larry   Completeness      100    100    100
#> 2 Larry   Validity          100      0     NA
reps <- c('Bob', 'Larry')
purrr::map_df(reps, ~ makeValidData(df, indexCol = 'myindex', filtRep = .x))
#> # A tibble: 4 × 5
#>   myindex metric       lastname gender norule
#>   <fct>   <chr>           <dbl>  <dbl>  <dbl>
#> 1 Bob     Completeness      100    100    100
#> 2 Bob     Validity          100    100     NA
#> 3 Larry   Completeness      100    100    100
#> 4 Larry   Validity          100      0     NA