Calculate time differences from 2 columns in a dataframe with various outputs

Compares 2 date columns in a dataframe. Wraps `base::difftime()` though with varying types of output based on arguments used. Supports custom functions to be applied to output of `base::difftime()`.

Usage

timeliness(
  df,
  fromCol,
  toCol,
  gpBy = NULL,
  FUNS = list(),
  colNames = list(),
  ...
)

Arguments

df

Dataframe A dataframe containing 2 columns of Date, POSIXct, or character coercible to date by `as.Date()`.

fromCol

Character Name of the column containing the start dates.

toCol

Character Name of the column containing the end dates.

gpBy

Character Vector of column names to group by. Ignored if `FUNS` is not used.

FUNS

List Optional aggregating functions (e.g., mean, median, std, sum...) or non-aggregating function to be applied to the output of `base::difftime()`. If grouping with `gpBy` is not used, aggregating functions return single value result column, and non-aggregating functions (e.g., function(x) x <= 7) return rowwise results. If grouping with `gpBy` is used, aggregating functions return a single value for each group, non-aggregating functions are ignored and a warning is displayed.

colNames

List List of characters to use as column names for columns created by functions in `FUNS`.

...

Additional arguments passed to `base::difftime()`

Value

Original dataframe with column `timeliness` containing output of `base::difftime()`. If argument `FUNS` is used without grouping columns in `gpBy`, `FUNS` output is propagated in added columns named from values in `colNames`. If `gpBy` is used and `FUNS` contains aggregating functions (e.g., mean), returns summary dataframe with grouping columns and columns defined by `FUNS` and `colNames`. Non-aggregating functions are ignored if `gpBy` is used.

Examples

dft <- data.frame(
'group1' = rep(c('A', 'B', "C"), 100),
'group2' = rep(c('F', 'U'), 150),
'd1' = c(seq.Date(as.Date('2023-01-01'), as.Date('2023-10-23'), 'day'), rep(NA_character_, 4))
) %>% 
dplyr::mutate(
    d2 = d1 + sample(1:100, 300, replace = TRUE)
)

# Will warn about NAs
out <- timeliness(dft, 'd1', 'd2', units = 'hours')
#> Warning: 4 rows ignored where date columns NA.
head(out)
#>   group1 group2         d1         d2 timeliness
#> 1      A      F 2023-01-01 2023-03-27 2040 hours
#> 2      B      U 2023-01-02 2023-01-24  528 hours
#> 3      C      F 2023-01-03 2023-01-24  504 hours
#> 4      A      U 2023-01-04 2023-03-28 1992 hours
#> 5      B      F 2023-01-05 2023-02-20 1104 hours
#> 6      C      U 2023-01-06 2023-01-20  336 hours
tail(out)
#>     group1 group2         d1         d2 timeliness
#> 295      A      F 2023-10-22 2024-01-05 1800 hours
#> 296      B      U 2023-10-23 2023-11-06  336 hours
#> 297      C      F       <NA>       <NA>   NA hours
#> 298      A      U       <NA>       <NA>   NA hours
#> 299      B      F       <NA>       <NA>   NA hours
#> 300      C      U       <NA>       <NA>   NA hours

# Removing NAs
d <- dft %>% 
dplyr::filter(., !is.na(d1))

out <- timeliness(d, 'd1', 'd2', units = 'hours')
head(out)
#>   group1 group2         d1         d2 timeliness
#> 1      A      F 2023-01-01 2023-03-27 2040 hours
#> 2      B      U 2023-01-02 2023-01-24  528 hours
#> 3      C      F 2023-01-03 2023-01-24  504 hours
#> 4      A      U 2023-01-04 2023-03-28 1992 hours
#> 5      B      F 2023-01-05 2023-02-20 1104 hours
#> 6      C      U 2023-01-06 2023-01-20  336 hours

suppressWarnings(
    out <- timeliness(dft, fromCol = 'd1', toCol = 'd2',
        FUNS = list(mean,median, function(x) x, function(x) x <= 7),
        colNames = list('mean_timeliness', 'median_timeliness', 
                'time_diff', 'lt_1_week'),
             units = 'days'
    )
)
head(out)
#>   group1 group2         d1         d2 mean_timeliness median_timeliness
#> 1      A      F 2023-01-01 2023-03-27   48.34797 days         47.5 days
#> 2      B      U 2023-01-02 2023-01-24   48.34797 days         47.5 days
#> 3      C      F 2023-01-03 2023-01-24   48.34797 days         47.5 days
#> 4      A      U 2023-01-04 2023-03-28   48.34797 days         47.5 days
#> 5      B      F 2023-01-05 2023-02-20   48.34797 days         47.5 days
#> 6      C      U 2023-01-06 2023-01-20   48.34797 days         47.5 days
#>   time_diff lt_1_week
#> 1   85 days     FALSE
#> 2   22 days     FALSE
#> 3   21 days     FALSE
#> 4   83 days     FALSE
#> 5   46 days     FALSE
#> 6   14 days     FALSE

# Using grouping with gpBy argument

out <- timeliness(d, fromCol = 'd1', toCol = 'd2', gpBy = c('group1', 'group2'),
    FUNS = list(mean, median),
    colNames = list('mean_timeliness', 'median_timeliness'),
    units = 'days'
)
head(out)
#>   group1 group2 mean_timeliness median_timeliness
#> 1      A      F   49.70000 days           49 days
#> 2      B      U   46.44000 days           44 days
#> 3      C      F   38.93878 days           34 days
#> 4      A      U   49.08163 days           45 days
#> 5      B      F   53.51020 days           55 days
#> 6      C      U   52.42857 days           55 days