US zipcodes database

A dataframe of US zipcodes and related geographic info

Usage

zipcodeDB

Format

A dataframe with 41877 rows and 24 variables

zipcode

US 5-digit zipcodes

zipcode_type

2010 State FIPS Code

major_city

Major city serving the ZIP code

post_office_city

Major city serving the ZIP code

common_city_list

List of common cities represented by the ZIP code

county

Name of county containing the ZIP code

state

Two-digit state code for ZIP code location

lat

Latitude of the centroid for the ZIP code

lng

Longitude of the centroid for the ZIP code

timezone

Timezone of the ZIP code

radius_in_miles

Radius of the ZIP code in miles

area_code_list

List of area codes for telephone numbers within this ZIP code

population

Total population of the ZIP code

population_density

Population density of the ZIP code (persons per square mile)

land_area_in_sqmi

Area of the land contained within the ZIP code in square miles

water_area_in_sqmi

Area of the waters contained within the ZIP code in square miles

housing_units

Number of housing units within the ZIP code

occupied_housing_units

Number of housing units within the ZIP code

median_home_value

Median home price within the ZIP code

bounds_west

Bounding box coordinates

bounds_east

Bounding box coordinates

bounds_north

Bounding box coordinates

bounds_south

Bounding box coordinates

Source

Taken from the `zipcodeR` package. That package gives new warnings in 2023 due to reliance on deprecated package `sp`. THis dataset is all we use in validation and so has been excerpted to remove dependency on the `zipcodeR` package. Raw data source for updates: https://github.com/MacHu-GWU/uszipcode-project/files/5183256/simple_db.log

zipcodeR package on GitHub https://github.com/gavinrozzi/zipcodeR