Edit Rules in a Validation Rules List

Generic to edit items to a S3 list or vector

editRules(x, ruleset = "default", newrules)

Arguments

x

List of class validRules containing named rules to edit

ruleset

Character Name of rule set to edit the rules to

newrules

List Named list of functions to replace old functions with. If newrules names are present, they are updated. If not present, they are ignored.

See also

Examples

n <- 100
myDF <- data.frame(
    'numbers' = sample(c(84121,85236,65987,NA_real_), n, replace = TRUE),
    'chars' = sample(c('word','sentence of words', 'Bob', '42', '1111'), n, replace = TRUE),
    'dates' = sample(seq.Date(as.Date('2021-01-01'), Sys.Date() + 100, 'days'), n, replace = TRUE)
)

# Hypothetical validation requirements ----
   ## Find missing numbers in "numbers"
       ## Find numeric characters in "chars"
       ## Find dates older than 2 years or any dates in the future in "dates"

# Set up named list of rules ----

rules <- list(
    'numbers' = function(x) is.na(as.numeric(x)),
    'chars' = function(x) grepl('[^0-9]', x),
    'dates' = function(x) {!is.na(as.POSIXct(x, format = "%m/%d/%Y")) & 
        as.POSIXct(x, format = "%m/%d/%Y") <= as.POSIXct(Sys.Date()) & 
        as.POSIXct(x, format = "%m/%d/%Y") > as.POSIXct(Sys.Date() - 2*365)}
)

# Create class `validRules` ruleset ----

myRules <- validRules(rules, ruleset = 'myRuleset')

validateDF(myDF, rulesObj = myRules, ruleset = 'myRuleset')
#> # A tibble: 1 × 3
#>   numbers chars dates
#>     <dbl> <dbl> <dbl>
#> 1      21    58    45

# Change some rules ----

myRules <- editRules(x = myRules, ruleset = 'myRuleset',
    newrules = list(
        'chars' = function(x) x == 'Bob',
        'notpresentrule' = function(x) x > 7000
    )                     
)
#> Ignoring unmatched rule names: notpresentrule

validateDF(myDF, rulesObj = myRules, ruleset = 'myRuleset')
#> # A tibble: 1 × 3
#>   numbers chars dates
#>     <dbl> <dbl> <dbl>
#> 1      21    19    45