Calculate time differences from 2 columns in a dataframe with various outputs

Compares 2 date columns in a dataframe. Wraps `base::difftime()` though with varying types of output based on arguments used. Supports custom functions to be applied to output of `base::difftime()`.

Usage

timeliness(
  df,
  fromCol,
  toCol,
  gpBy = NULL,
  FUNS = list(),
  colNames = list(),
  ...
)

Arguments

df

Dataframe A dataframe containing 2 columns of Date, POSIXct, or character coercible to date by `as.Date()`.

fromCol

Character Name of the column containing the start dates.

toCol

Character Name of the column containing the end dates.

gpBy

Character Vector of column names to group by. Ignored if `FUNS` is not used.

FUNS

List Optional aggregating functions (e.g., mean, median, std, sum...) or non-aggregating function to be applied to the output of `base::difftime()`. If grouping with `gpBy` is not used, aggregating functions return single value result column, and non-aggregating functions (e.g., function(x) x <= 7) return rowwise results. If grouping with `gpBy` is used, aggregating functions return a single value for each group, non-aggregating functions are ignored and a warning is displayed.

colNames

List List of characters to use as column names for columns created by functions in `FUNS`.

...

Additional arguments passed to `base::difftime()`

Value

Original dataframe with column `timeliness` containing output of `base::difftime()`. If argument `FUNS` is used without grouping columns in `gpBy`, `FUNS` output is propagated in added columns named from values in `colNames`. If `gpBy` is used and `FUNS` contains aggregating functions (e.g., mean), returns summary dataframe with grouping columns and columns defined by `FUNS` and `colNames`. Non-aggregating functions are ignored if `gpBy` is used.

Examples

dft <- data.frame(
'group1' = rep(c('A', 'B', "C"), 100),
'group2' = rep(c('F', 'U'), 150),
'd1' = c(seq.Date(as.Date('2023-01-01'), as.Date('2023-10-23'), 'day'), rep(NA_character_, 4))
) %>% 
dplyr::mutate(
    d2 = d1 + sample(1:100, 300, replace = TRUE)
)

# Will warn about NAs
out <- timeliness(dft, 'd1', 'd2', units = 'hours')
#> Warning: 4 rows ignored where date columns NA.
head(out)
#>   group1 group2         d1         d2 timeliness
#> 1      A      F 2023-01-01 2023-01-07  144 hours
#> 2      B      U 2023-01-02 2023-01-07  120 hours
#> 3      C      F 2023-01-03 2023-01-15  288 hours
#> 4      A      U 2023-01-04 2023-01-12  192 hours
#> 5      B      F 2023-01-05 2023-01-16  264 hours
#> 6      C      U 2023-01-06 2023-02-07  768 hours
tail(out)
#>     group1 group2         d1         d2 timeliness
#> 295      A      F 2023-10-22 2023-11-09  432 hours
#> 296      B      U 2023-10-23 2023-11-06  336 hours
#> 297      C      F       <NA>       <NA>   NA hours
#> 298      A      U       <NA>       <NA>   NA hours
#> 299      B      F       <NA>       <NA>   NA hours
#> 300      C      U       <NA>       <NA>   NA hours

# Removing NAs
d <- dft %>% 
dplyr::filter(., !is.na(d1))

out <- timeliness(d, 'd1', 'd2', units = 'hours')
head(out)
#>   group1 group2         d1         d2 timeliness
#> 1      A      F 2023-01-01 2023-01-07  144 hours
#> 2      B      U 2023-01-02 2023-01-07  120 hours
#> 3      C      F 2023-01-03 2023-01-15  288 hours
#> 4      A      U 2023-01-04 2023-01-12  192 hours
#> 5      B      F 2023-01-05 2023-01-16  264 hours
#> 6      C      U 2023-01-06 2023-02-07  768 hours

suppressWarnings(
    out <- timeliness(dft, fromCol = 'd1', toCol = 'd2',
        FUNS = list(mean,median, function(x) x, function(x) x <= 7),
        colNames = list('mean_timeliness', 'median_timeliness', 
                'time_diff', 'lt_1_week'),
             units = 'days'
    )
)
head(out)
#>   group1 group2         d1         d2 mean_timeliness median_timeliness
#> 1      A      F 2023-01-01 2023-01-07   48.69595 days           48 days
#> 2      B      U 2023-01-02 2023-01-07   48.69595 days           48 days
#> 3      C      F 2023-01-03 2023-01-15   48.69595 days           48 days
#> 4      A      U 2023-01-04 2023-01-12   48.69595 days           48 days
#> 5      B      F 2023-01-05 2023-01-16   48.69595 days           48 days
#> 6      C      U 2023-01-06 2023-02-07   48.69595 days           48 days
#>   time_diff lt_1_week
#> 1    6 days      TRUE
#> 2    5 days      TRUE
#> 3   12 days     FALSE
#> 4    8 days     FALSE
#> 5   11 days     FALSE
#> 6   32 days     FALSE

# Using grouping with gpBy argument

out <- timeliness(d, fromCol = 'd1', toCol = 'd2', gpBy = c('group1', 'group2'),
    FUNS = list(mean, median),
    colNames = list('mean_timeliness', 'median_timeliness'),
    units = 'days'
)
head(out)
#>   group1 group2 mean_timeliness median_timeliness
#> 1      A      F   46.22000 days           44 days
#> 2      B      U   40.22000 days           35 days
#> 3      C      F   49.14286 days           45 days
#> 4      A      U   52.93878 days           55 days
#> 5      B      F   54.34694 days           56 days
#> 6      C      U   49.53061 days           48 days