Organize and access validation rules from a list of class `validRules`

List of validation rule functions for EDX validation use. THe default ruelset is from the EMSA QA2 application. If different rules are needed, create your own named list of rules and call validRules() to create the rules object. See examples.

Usage

validRules(x = list(), ruleset = "default")

Arguments

x

Named List or vector of validation rule functions. Names should correspond to column names in a dataframe being validated

ruleset

Character Grouping ruleset group name to assign the list in x. Multiple rulesets can be contained in a validRUles object. Add new rulesets with `addRules()`. See `?addRules`

Value

List of validation functions named for the column name they validate

See also

Examples

df <- data.frame("lastname" = c("AAA", "BBB"))

# Use default rules for EMSA QA2 Datasource
names(validRules())
#> [1] "default"
names(validRules()[['default']])
#>  [1] "firstname"         "lastname"          "dob"              
#>  [4] "street"            "city"              "state"            
#>  [7] "zipcode"           "phone"             "race"             
#> [10] "ethnicity"         "gender"            "pregnancy"        
#> [13] "reportingagency"   "diagnosticstreet"  "diagnosticcity"   
#> [16] "diagnosticstate"   "diagnosticzip"     "facility_street"  
#> [19] "facility_city"     "facility_state"    "facility_zip"     
#> [22] "clinician"         "clinicianphone"    "providerphone"    
#> [25] "visit_date"        "performinglab"     "testresult"       
#> [28] "datetested"        "collectiondate"    "datereported"     
#> [31] "localTestValid"    "clinicianfacility" "discharge_date"   
validRules()[["default"]][["lastname"]]
#> function (x) 
#> grepl("^[a-zA-Z]+[a-zA-Z,\\',\\- ]*[a-zA-Z ]*?$", x, perl = TRUE)
#> <bytecode: 0x55e988721c00>
#> <environment: 0x55e9907321e8>
validRules()[["default"]][["lastname"]](df$lastname)
#> [1] TRUE TRUE

# Create custom rule lists
myrules <- list(
    'columnname1' = function(x) x + 1,
    'columnname2' = function(x) x - 1
)

validRules(myrules)
#> # A tibble: 2 × 3
#>   Ruleset Name        Rule             
#>   <chr>   <chr>       <chr>            
#> 1 default columnname1 function(x) x + 1
#> 2 default columnname2 function(x) x - 1

myRuleSet <- validRules(myrules, ruleset = 'myRulesetName')

myrules2 <- list(
    "ruleset1" = list(
        'columnname1' = function(x) x + 1,
        'columnname2' = function(x) x - 1
    ),
    
    "ruleset2" = list(
        'columnname1' = function(x) x + 1,
        'columnname2' = function(x) x - 1
    )
)

validRules(myrules2)
#> # A tibble: 4 × 3
#>   Ruleset  Name        Rule             
#>   <chr>    <chr>       <chr>            
#> 1 ruleset1 columnname1 function(x) x + 1
#> 2 ruleset1 columnname2 function(x) x - 1
#> 3 ruleset2 columnname1 function(x) x + 1
#> 4 ruleset2 columnname2 function(x) x - 1
validRules(myrules2) %>% addRules(myRuleSet, ruleset = 'myRulesetName')
#> # A tibble: 6 × 3
#>   Ruleset       Name        Rule             
#>   <chr>         <chr>       <chr>            
#> 1 ruleset1      columnname1 function(x) x + 1
#> 2 ruleset1      columnname2 function(x) x - 1
#> 3 ruleset2      columnname1 function(x) x + 1
#> 4 ruleset2      columnname2 function(x) x - 1
#> 5 myRulesetName columnname1 function(x) x + 1
#> 6 myRulesetName columnname2 function(x) x - 1