Apply validateColumn and validRUles across a dataframe

Runs validation rules on all columns of dataframe df by column name. Returns NA if no validation rule exists

Usage

validateDF(df, rulesObj = NULL, ruleset = "default")

Arguments

df

A dataframe whose column values are subjected to validation rules via switch()

rulesObj

List Named list of validation functions

ruleset

Character Grouping name of custom ruleset.

Value

Single row dataframe with validation results and same columns as df

See also

Examples

df <- data.frame('lastname' = c('AAA','BBB','CCC'),
'gender' = c('Male','Female','M A L E'), 'norule' = c('A','B','C'))
validateDF(df)
#> # A tibble: 1 × 3
#>   lastname gender norule
#>      <dbl>  <dbl>  <dbl>
#> 1      100   66.7     NA
rulesObj <- validRules()
rulesObj <- rulesObj %>% addRules(list('norule' = function(x) x =='B'))
validateDF(df, rulesObj = rulesObj)
#> # A tibble: 1 × 3
#>   lastname gender norule
#>      <dbl>  <dbl>  <dbl>
#> 1      100   66.7   33.3