Validate multiple columns in each row of a dataframe

Identify rows in a dataframe where specified columns are all valid.

Usage

validateRows(df, cols = list(), rulesObj = NULL, ruleset = "default")

Arguments

df

dataframe whose column values are subjected to validation rules in combinations

cols

Vector of column names to validate

rulesObj

List Named list of validation functions

ruleset

Character Grouping name of custom ruleset.

Value

Returns original dataframe with flag column `valid_rows` flags `TRUE` where all specified columns return TRUE for their respective validation rules, else `FALSE`. Default behavior when `rulesObj` is not supplied returns flag column `valid_rows` flags `TRUE` when columns in `cols` satisfy `complete.cases()`. If no `cols` are supplied, all columns are included in the evaluation.

See also

Examples

set.seed(42)
df <- data.frame(
    'c1' = seq.Date(as.Date('1974-06-22'), as.Date('2000-06-24'), 'day'),
    'c2' = sample(100:10000, 9500, replace = TRUE),
    'c3' = sample(c('misfit', 'toys', 'island', 'rich', NA_character_), 9500, replace = TRUE),
    'c4' = sample(zipcodes, 9500, replace = TRUE),
    'c5' = sample(zipcodeDB$median_home_value, 9500, replace = TRUE)
)

rulesObj <- validRules(
    list(
        'c1' = function(x) !is.na(x) & is.Date(x),
        'c2' = function(x) x >= 500,
        'c3' = function(x) x %in% c('misfit', 'toys', 'island'),
        'c4' = function(x) nchar(x) == 5,
        'c5' = function(x) !is.na(x) & is.numeric(x)
    ),
    ruleset = 'test'
)

# Defaults to `complete.cases()` if `rulesObj` is not supplied

validated_df <- validateRows(df)
head(validated_df)
#>           c1   c2   c3    c4     c5 valid_rows
#> 1 1974-06-22 2468 toys 39745     NA      FALSE
#> 2 1974-06-23 5372 <NA> 72320 165000      FALSE
#> 3 1974-06-24 9389 toys 10543 147500       TRUE
#> 4 1974-06-25 1351 toys 12242 140700       TRUE
#> 5 1974-06-26 8925 toys 70093 177600       TRUE
#> 6 1974-06-27  455 <NA> 45145     NA      FALSE

# Defaults to `complete.cases()` for supplied `cols` if `rulesObj` is not supplied

validated_df <- validateRows(df, cols = c('c4', 'c5'))
head(validated_df)
#>           c1   c2   c3    c4     c5 valid_rows
#> 1 1974-06-22 2468 toys 39745     NA      FALSE
#> 2 1974-06-23 5372 <NA> 72320 165000       TRUE
#> 3 1974-06-24 9389 toys 10543 147500       TRUE
#> 4 1974-06-25 1351 toys 12242 140700       TRUE
#> 5 1974-06-26 8925 toys 70093 177600       TRUE
#> 6 1974-06-27  455 <NA> 45145     NA      FALSE

# Returns rows where all `cols` return TRUE for rules supplied in `rulesObj`

validated_df <- validateRows(
    df, 
    cols = c('c3', 'c4', 'c5'), 
    rulesObj = rulesObj, 
    ruleset = 'test'
)

head(validated_df, 20)
#>            c1   c2     c3    c4     c5 valid_rows
#> 1  1974-06-22 2468   toys 39745     NA      FALSE
#> 2  1974-06-23 5372   <NA> 72320 165000      FALSE
#> 3  1974-06-24 9389   toys 10543 147500       TRUE
#> 4  1974-06-25 1351   toys 12242 140700       TRUE
#> 5  1974-06-26 8925   toys 70093 177600       TRUE
#> 6  1974-06-27  455   <NA> 45145     NA      FALSE
#> 7  1974-06-28 7799 misfit 88572     NA      FALSE
#> 8  1974-06-29 4053 misfit 37340 144500       TRUE
#> 9  1974-06-30 9190   rich 66036 439800      FALSE
#> 10 1974-07-01 5502   rich 72003  96000      FALSE
#> 11 1974-07-02 1031   toys 65759     NA      FALSE
#> 12 1974-07-03 9288   <NA> 69145  67100      FALSE
#> 13 1974-07-04 5736   toys 59253     NA      FALSE
#> 14 1974-07-05 4101   <NA> 60645  76200      FALSE
#> 15 1974-07-06 9151   <NA> 92318     NA      FALSE
#> 16 1974-07-07  358   toys 84654     NA      FALSE
#> 17 1974-07-08 5533   toys 59064     NA      FALSE
#> 18 1974-07-09  580   <NA> 85013     NA      FALSE
#> 19 1974-07-10 7425   toys 21001  98300       TRUE
#> 20 1974-07-11 8590 island 33019  79600       TRUE